Addressing comments: - Rephrasing one paragraph - Adding a `remove` to the react component to avoid multiple drawings - Renaming setup to chart-component - Removing react-components as it's no longer used
